requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.5.1
joblib>=1.4.2
pyyaml>=6.0.2
# annoy==1.17.3